Validate StageProp, SequenceChannel and PlaylistSequence UUIDs on server (#28)

+++ persistence/src/main/java/io/sparkled/persistence/stage/impl/query/SaveStagePropsQuery.java
@@ -3,6 +3,7 @@ package io.sparkled.persistence.stage.impl.query;
 import io.sparkled.model.entity.Stage;
 import io.sparkled.model.entity.StageProp;
 import io.sparkled.model.validator.StagePropValidator;
+import io.sparkled.model.validator.exception.EntityValidationException;
 import io.sparkled.persistence.PersistenceQuery;
 import io.sparkled.persistence.QueryFactory;
 import org.slf4j.Logger;
@@ -29,14 +30,30 @@ public class SaveStagePropsQuery implements PersistenceQuery<Void> {
     @Override
     public Void perform(QueryFactory queryFactory) {
         StagePropValidator stagePropValidator = new StagePropValidator();
+        stageProps.forEach(sp -> sp.setStageId(stage.getId()));
         stageProps.forEach(stagePropValidator::validate);
 
-        EntityManager entityManager = queryFactory.getEntityManager();
-        stageProps.forEach(entityManager::merge);
-        logger.info("Saved {} stage prop(s) for stage {}.", stageProps.size(), stage.getId());
+        if (uuidAlreadyInUse(queryFactory)) {
+            throw new EntityValidationException("Stage prop already exists on another stage.");
+        } else {
+            EntityManager entityManager = queryFactory.getEntityManager();
+            stageProps.forEach(entityManager::merge);
+            logger.info("Saved {} stage prop(s) for stage {}.", stageProps.size(), stage.getId());
 
-        deleteRemovedStageProps(queryFactory);
-        return null;
+            deleteRemovedStageProps(queryFactory);
+            return null;
+        }
+    }
+
+    private boolean uuidAlreadyInUse(QueryFactory queryFactory) {
+        List<UUID> uuidsToCheck = stageProps.stream().map(StageProp::getUuid).collect(toList());
+        uuidsToCheck = uuidsToCheck.isEmpty() ? noUuids : uuidsToCheck;
+
+        long uuidsInUse = queryFactory.select(qStageProp)
+                .from(qStageProp)
+                .where(qStageProp.stageId.ne(stage.getId()).and(qStageProp.uuid.in(uuidsToCheck)))
+                .fetchCount();
+        return uuidsInUse > 0;
     }
 
     private void deleteRemovedStageProps(QueryFactory queryFactory) {
